third_party/blink/web_tests/dom/document_type/remove.html

<!DOCTYPE html>
<title>This tests the DOM 4 remove method on a DocumentType node</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var doctype = document.doctype;
assert_equals(document.childNodes.length, 2);
doctype.remove();
assert_equals(document.childNodes.length, 1);
doctype.remove();
assert_equals(document.childNodes.length, 1);
done();
</script>
